conceptos desde creacion de empresa

+++ database/migrations/2020_08_05_141624_create_empresas_table.php
@@ -37,7 +37,7 @@ class CreateEmpresasTable extends Migration
             $table->string('tipoPeriodo',10);
             $table->date('inicioPeriodo');
             $table->string('region',50);
-            $table->double('primaRiesgo');
+            $table->float('primaRiesgo');
             $table->double('porcentajeAhorro');
             $table->timestamps();
         });
